+++ packages/common/evmUtils/src/operations/token/index.ts
@@ -1,4 +1,5 @@
 export * from './getErc20BurnsOperation';
+export * from './getErc20TransfersOperation';
 export * from './getErc20MintsOperation';
 export * from './getTokenAllowanceOperation';
 export * from './getTokenMetadataBySymbolOperation';